issues/30: break ReloadButtonCoordinator into a new component
